postgres refactor
